クリアタイムを保存してクリアシーンに描画させた

+++ GameData.cpp
@@ -2,6 +2,6 @@
 
 // 実体定義（ここだけ！）
 int g_score = 0;
-int g_time = 0;
-int g_clearTime = 0;
+float g_time = 0.0f;
+float g_clearTime = 0.0f;
 bool g_isCleared = false;
