+++ GameData.h
@@ -2,6 +2,6 @@
 extern int g_score;
 
 // ƒ^ƒCƒ€
-extern int g_time;
-extern int g_clearTime;
+extern float g_time;
+extern float g_clearTime;
 extern bool g_isCleared;
